+++ src/components/mypage/inside/roadmap/Roadmap.js
@@ -1,39 +1,9 @@
-import Button from "../../../Button/Button";
 import "./Roadmap.css"
 import {useEffect, useState} from "react";
-import { useParams, useNavigate } from "react-router-dom";
 import axios from "axios";
-import {useAuth} from "../../../Login/AuthContext";
 
-const Roadmap = ({ setManage }) => {
-    const [allRoadmaps, setAllRoadmaps] = useState([]);
-    const [selectedRoadmapId, setSelectedRoadmapId] = useState("");
+const Roadmap = ({ selectedRoadmapId }) => {
     const [roadmapDetail, setRoadmapDetail] = useState(null);
-    const { user } = useAuth();
-    const { roadmapId } = useParams();
-    const navigate = useNavigate();
-
-    useEffect(() => {
-        axios.get(`http://localhost:8080/board/roadmaps/${user.memId}`, {
-            headers: {
-                'Accept': 'application/json'
-            }
-        })
-            .then(response => {
-                const roadmapData = response.data;
-
-                setAllRoadmaps(roadmapData || []);
-                if (roadmapData && roadmapData.length > 0) {
-                    const firstRoadmapId = roadmapId || roadmapData[0].id;
-                    setSelectedRoadmapId(firstRoadmapId);
-                    navigate(`/mypage/roadmap/${firstRoadmapId}`);
-                }
-            })
-
-            .catch(error => {
-                console.log('로드맵 리스트 조회 에러', error)
-            })
-    }, []);
 
     useEffect(() => {
         if (selectedRoadmapId) {
@@ -52,24 +22,8 @@ const Roadmap = ({ setManage }) => {
             });
     };
 
-    const handleRoadmapChange = (e) => {
-        const newRoadmapId = e.target.value;
-        navigate(`/mypage/roadmap/${newRoadmapId}`);
-        setSelectedRoadmapId(newRoadmapId); // 선택한 로드맵 ID 저장
-    };
-
     return (
         <div>
-            <div>
-                <select value={selectedRoadmapId} onChange={handleRoadmapChange}>
-                    {allRoadmaps.map(roadmap => (
-                        <option key={roadmap.id} value={roadmap.id}>
-                            {roadmap.title}
-                        </option>
-                    ))}
-                </select>
-                <Button label="로드맵 관리" className="" onClick={setManage}/>
-            </div>
             <div className="all-contatiner">
                 {roadmapDetail && (
                     <div>
